+++ src/main/java/net/id107/flexfov/gui/advanced/FisheyeGui.java
@@ -76,7 +76,7 @@ public class FisheyeGui extends AdvancedGui {
 		DoubleOption FOV = new DoubleOption("fisheyeFov", 0, fovSliderLimit, 1,
 				(gameOptions) -> {return Math.min(finalSliderLimit, Projection.getProjection().getFovX());},
 				(gameOptions, number) -> {Projection.fov = number; ConfigManager.saveConfig();},
-				(gameOptions, doubleOption) -> {return new LiteralText("FOV: " + Math.min(finalSliderLimit, Projection.getProjection().getFovX()));});
+				(gameOptions, doubleOption) -> {return new LiteralText("FOV: " + (int)Math.min(finalSliderLimit, Projection.getProjection().getFovX()));});
 		addButton(FOV.createButton(client.options, width / 2 - 180, height / 6 + 132, fovSliderLimit));
 		
 		addButton(new ButtonWidget(width / 2 - 155, height / 6 + 84, 150, 20,
